added vulkan backend

+++ src/core/clock.hpp
@@ -9,26 +9,25 @@ public:
   void reset() {
     startTime = std::chrono::high_resolution_clock::now();
     lastTime = startTime;
-    deltaTime = 0.0f;
+    deltaTime = 0.0;
   }
 
   void update() {
     auto currentTime = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<float> elapsed = currentTime - lastTime;
+    std::chrono::duration<double> elapsed = currentTime - lastTime;
     deltaTime = elapsed.count();
     lastTime = currentTime;
   }
 
-  float getDeltaTime() const { return deltaTime; }
+  double getDeltaTime() const { return deltaTime; }
 
-  float getElapsedTime() const {
-    std::chrono::duration<float> elapsed =
-        std::chrono::high_resolution_clock::now() - startTime;
+  double getElapsedTime() const {
+    std::chrono::duration<double> elapsed = std::chrono::high_resolution_clock::now() - startTime;
     return elapsed.count();
   }
 
 private:
   std::chrono::high_resolution_clock::time_point startTime;
   std::chrono::high_resolution_clock::time_point lastTime;
-  float deltaTime;
+  double deltaTime;
 };
